backend/pyproject.toml
[project]
name = "TechTix"
version = "0.1.0"
description = "Test"
requires-python = ">=3.11,<3.12"

dependencies = [
    "pynamodb==6.1.0",
    "ulid==1.1",
    "fastapi==0.96.0",
    "fastapi-cloudauth==0.4.3",
    "mangum==0.15.0",
    "lambda-warmer-py==0.6.0",
    "typing-extensions==4.8.0",
    "pydantic[email]>=1.10.0,<2.0.0",
    "lambda-decorators==0.6.0",
    "python-dateutil==2.9.0.post0",
    "requests==2.32.3",
    "pytz==2024.2",
    "polars>=1.33.1",
    "openpyxl>=3.1.5",
    "pandas>=2.3.2",
    "pillow>=11.3.0",
]

[dependency-groups]
dev = [
    "commitizen>=4.8.3",
    "boto3==1.40.11",
    "python-dotenv==1.0.0",
    "uvicorn==0.22.0",
    "pre-commit==3.3.3",
    "pytest==7.1.1",
    "pytest-cov==3.0.0",
    "pytest-mock==3.10.0",
    "moto==3.1.1",
    "pylint==2.13.8",
    "pylint-pydantic==0.1.8",
    "httpx==0.23.3",
    "ruff==0.12.2",
    "mypy>=1.17.0",
    "detect-secrets>=1.5.0",
    "flake8>=7.3.0",
    "black>=25.1.0",
]

[tool.mypy]
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
allow_untyped_defs = false
disallow_untyped_calls = true
no_implicit_optional = true
check_untyped_defs = true
strict = false
exclude = [
  '^.venv/',
  '.*/site-packages/',
  '^.serverless/',
]

[tool.ruff]
line-length = 120
indent-width = 4
target-version = "py311"
exclude = [
".bzr",
".direnv",
".eggs",
".git",
".git-rewrite",
".hg",
".ipynb_checkpoints",
".mypy_cache",
".nox",
".pants.d",
".pyenv",
".pytest_cache",
".pytype",
".ruff_cache",
".svn",
".tox",
".venv",
".vscode",
"pypackages",
"_build",
"buck-out",
"build",
"dist",
"node_modules",
"site-packages",
"venv",
".serverless"
]

[tool.ruff.lint]
ignore = [
"E203", "E266", "E501", "F403", "F401", "E402", "F841", "C901", "F722", "F405", "F811"
]
select = ["E4", "E7", "E9", "F"]
fixable = ["ALL"]
unfixable = []
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = false
docstring-code-line-length = "dynamic"

[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "$version"
version_scheme = "pep440"
version_provider = "uv"
major_version_zero = true
